+++ src/CloudMigrator.Cli/Program.cs
@@ -1,2 +1,9 @@
-﻿// See https://aka.ms/new-console-template for more information
-Console.WriteLine("Hello, World!");
+﻿using System.CommandLine;
+using CloudMigrator.Cli.Commands;
+
+var rootCmd = new RootCommand("CloudMigrator - OneDrive から SharePoint へのファイル移行ツール");
+rootCmd.Add(TransferCommand.Build());
+rootCmd.Add(RebuildSkipListCommand.Build());
+
+return await rootCmd.Parse(args).InvokeAsync(new InvocationConfiguration(), CancellationToken.None);
+